+++ lib/halitosis.rb
@@ -40,7 +40,6 @@ module Halitosis
     base.include Meta
     base.include Permissions
     base.include Preloadable
-    base.include Relationships
     base.include RootLinks
     base.include RootMeta
     base.include RootPermissions
@@ -95,7 +94,7 @@ require_relative "halitosis/links"
 require_relative "halitosis/meta"
 require_relative "halitosis/permissions"
 require_relative "halitosis/preloadable"
-require_relative "halitosis/relationships"
+require_relative "halitosis/resource_relationships"
 require_relative "halitosis/root_links"
 require_relative "halitosis/root_meta"
 require_relative "halitosis/root_permissions"
