+++ lib/halitosis/resource.rb
@@ -11,6 +11,7 @@ module Halitosis
 
       base.extend ClassMethods
 
+      base.send :include, ResourceRelationships
       base.send :include, InstanceMethods
 
       base.send :attr_reader, :resource
