+++ spec/integrations/include_options_spec.rb
@@ -3,7 +3,7 @@ RSpec.describe "Include Options" do
     let(:klass) do
       Class.new do
         include Halitosis
-        include Halitosis::Relationships
+        include Halitosis::ResourceRelationships
 
         attribute(:verify_depth) { |ctx| ctx.depth }
         attribute(:verify_include) { |ctx| ctx.include_options.keys }
@@ -225,7 +225,7 @@ RSpec.describe "Include Options" do
           [
             Class.new do
               include Halitosis
-              include Halitosis::Relationships
+              include Halitosis::ResourceRelationships
 
               attribute(:verify_depth) { |ctx| ctx.depth }
               attribute(:verify_include) { |ctx| ctx.include_options.keys }
